+++ ui/cypress/support/commands.ts
@@ -28,6 +28,37 @@ export const createDashboard = (
   })
 }
 
+export const createDashboardTemplate = (
+  orgID?: string,
+  name: string = 'Bashboard'
+): Cypress.Chainable<Cypress.Response> => {
+  return cy.request({
+    method: 'POST',
+    url: '/api/v2/documents/templates',
+    body: {
+      content: {
+        data: {
+          attributes: {name, description: ''},
+          relationships: {
+            label: {data: []},
+            cell: {data: []},
+            variable: {data: []},
+          },
+          type: 'dashboard',
+        },
+        included: [],
+      },
+      labels: [],
+      meta: {
+        description: `template created from dashboard: ${name}`,
+        version: '1',
+        name: `${name}-Template`,
+      },
+      orgID,
+    },
+  })
+}
+
 export const createOrg = (): Cypress.Chainable<Cypress.Response> => {
   return cy.request({
     method: 'POST',
@@ -270,6 +301,7 @@ Cypress.Commands.add('setupUser', setupUser)
 
 // dashboards
 Cypress.Commands.add('createDashboard', createDashboard)
+Cypress.Commands.add('createDashboardTemplate', createDashboardTemplate)
 
 // orgs
 Cypress.Commands.add('createOrg', createOrg)