+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## v2.0.0-alpha.8 [unreleased]
 
 ### Features
+
 1. [13024](https://github.com/influxdata/influxdb/pull/13024): Add the ability to edit token's description
+1. [13078](https://github.com/influxdata/influxdb/pull/13078): Add the option to create a Dashboard from a Template.
 
 ### Bug Fixes